Unify behavior when nothing is to be done
This patch lets commands uniformly check the `Inventory` for pending
operations before trying to display a diff and asking for confirmation,
avoiding pointless empty lines and "questions about nothing".

`Inventory.modify_asset`, `Inventory.rename_asset` and
`Inventory.rename_directory` can potentially raise a `NoopError`,
allowing the caller to decide whether this should lead to failure or not
(following a result-oriented approach where a call to these methods is
not lookad at as "do this" but as "make sure this the state afterwards").