Add fixture for ui setup in tests
This provides a fixture setting up `onyo.lib.ui`, that is run for every
test. Settings for `yes`, `quiet` and `debug` are read from a new test
marker 'ui' if given. Otherwise set to defaults (`False` for all of
them). The marker takes a dictionary, allowing to enhance this later on
for other future features of the `UI` object (like using the rich
package or redirecting logging or whatever else).