commands.py: pass an Inventory instead of OnyoRepo to command-functions
Close #467
Instead of passing sometimes an Inventory and sometimes an OnyoRepo,
this changes the command-functions in commands.py and the variables
in onyo/commands/*.py to always be an Inventory object, so as to make
the functions more consistent.

Functions still to be refactored (fsck and unset) were not changed.