Rework `onyo_new`
This patch addresses several issues with `onyo_new`. Most importantly,
it rectifies the order of the generation of potential 'faux' serials,
the asset's name, the initialization with a template, and the validation
of the resulting, to-be-created assets. This was inconsistently done
depending on whether `edit` was `True`, templates used, etc. (and
inconsistent across onyo's `new`, `edit`, and `set`)

The overall approach is to move these functionalities into the
`Inventory` methods that register operations (in case of `new`, that's
`Inventory.add_asset`. Now, `new` would try to add an asset and deal
with potential exceptions. This includes the possibility to "unregister"
operations after a (partially) successful trial to add them. This
approach is intended for all commands. An additional aspect of it is the
envisioned third-party use of the python interface. Since these
`Inventory` methods can be used directly instead of via an onyo command,
the validation (and generation) needs to happen within. Which implies
that command implementations that do the same thing before even trying
to execute, would lead to duplicate testing of inputs.

Because `Inventory.add_asset` now generates the asset itself, it does
not respect an incoming 'path' key anymore, but only a 'directory' key
assigning the path itself as a result of the name generation.

`onyo_new(edit=True, ...` now shows a diff of the operations implied
after a successful edit, allowing a user to keep changing the asset
based on an evaluation of the effective changes (generated stuff + its
validation), rather than only seeing this when all assets are edited
with no way to go back.
As a result, the confirmation of a total diff at the end of the command
is disabled when `edit=True` was given, since every effective diff per
asset was confirmed already.