RF: Change inventory asset queries
This changes the behavior of `Inventory.get_assets_by_query` in
preparation of reworking the get command.
- Do not account for the `keys` option anymore. This option is about
  shaping the output not about retrieving matches.
- Take callables as filters instead of `Filter` instances.
  This enables callers of that function to utilize whatever mechanism
  they like to define matches.
- Remove outdated notion of pseudo keys being encoded in filenames.

Regarding type annotation `Generator | filter`:

For some reason pyre does not accept a `Generator` annotation, when it
could be the result of the builtin `filter`. Have not figured out, how
to specify a generator, that includes both. There should be a better
solution to that.