BF: Consider committed files only
This patch changes the `GitRepo.get_subtrees` method to use `git
ls-tree` instead of `git ls-files`, in order to consistently only
consider committed files instead of what's in the worktree.
This was missed when addressing issue #502.

There's a corner case when a repository doesn't have a commit and
therefore no HEAD. Returning an empty list is just fine in that case.
Therefore, catch the error and pretend the output was empty.