Adjust tests for commands not running on dirty worktree
Removes untracked content from inventory fixture. This wasn't actually
used in tests and isn't useful in this from anymore. While testing
functions/commands on a dirty worktree is desirable, the currently only
fixture delivering an `Inventory` isn't usefu if it always comes with a
dirty worktree.

For the same reason, change the way the test for editor config
precedence works to commit changes to ``.onyo/config`` in order to allow
subsequent onyo command execution.