make zsh shell_completion static
This removes the logic to parse argparse (imperfectly) to generate
a ZSH completion script, and instead simply uses a static file.

The auto-generation idea was nice, and it was remarkably complete, but
the resulting logic was caught between argparse's awkward, undefined API
and ZSH's insanity. The results were not pretty.

A static file will allow all ZSH features to be used without gymnastics
to work around arparse.

The `shell-completion` command is kept, because ZSH is still not
XDG-aware, and has no counterpart to bash-completion's
~/.local/share/bash-completion/completions

The static file is the direct output of `onyo shell-completion` with no
modification of the content.