BF: `rm` handles duplication and fails on inventory root
- Fixes `onyo rm` to account for already pending removals. This should
  be done more consistently (see #546).
- Fixes `onyo rm` to fail when trying to remove the inventory's root
  dir.

Note, that changes in `Inventory.remove_directory` are mostly about
recursive removal of directories. For example, changing a condition to
recursively call itself on `path.is_dir` rather than considering
`is_inventory_dir` is done, because the latter, more restrictive
condition would be checked again during recursion anyway.

Refactoring of the "dummy operation" for generic file removal is done,
so it can be easily accounted for by `Inventory._get_pending_removals`,
which would be harder to do if it remained that unnamed ad-hoc
operation.

Lastly, note that the unskipped test's referral to #309 was wrong to
begin with - #519 is what it was about.

(Closes #519)
(Closes #520)