`set`: Remove matching and add asset dirs support
This patch removes functionality from `set` that allowed to query for
assets. That way, `set` does not have to deal with paths being either
assets or dirs, but only ever assets. This disambiguates what `set` is
supposed to do when being pointed to an asset dir.

At the same time, the demo generation script proves, that one can use
`get` for querying and feed the result into `set`, which seems like a
cleaner approach.