`rm`: Add support for asset dirs
This adds option(s) to the `rm` command, that allow to switch modes of
operation. It either removes assets only, in which case it fails on
regular directories and turns asset dirs into regular dirs (removing the
asset "aspect" of them), or it removes directories only, in which case
it fails on assets and turns asset dirs into asset files instead
(removing the directory "aspect"). Last mode is the default and matches
previous behavior: Remove everything in any case.

For python interface, this is realized as a `mode` switch with three
valid values and for the CLI the mutually exclusive, optional switches
`-a|--asset` and `-d|--dir` are added instead.