DOC/BF: python -> python3
Inside a virtual-env `python` maps to `python3`. However, outside of
that, convention is that `python` is reserved for `python2.7` (which is
long since deprecated.

This switched `python` and `pip` to explicitly specify the version 3
variants. Also, the instructions for Ubuntu/Debian are updated to
install pip, since it is not part of the main python3 package.

fix #569