Machine-readable `get` quotes values
This is changing the output of `onyo_get` when used in machine-readable
mode to quote the values it prints. Hence, whitespaces in values aren't
an issue (especially when piping `onyo get` into something else.

Demo generation script is adjusted accordingly, not requiring the `-d
"\n"` option for `xargs` anymore, when reading `onyo get`'s output to
construct the `onyo set` call.