NF/DOC: generate Python API docs
This auto-generates the Python API docs for Onyo.

It limits auto-generation to `onyo` and `onyo.lib`. Commands are
excluded because they're already included in "Command Line Reference".
Various internal helpers are excluded to reduce irrelevant mess.

The RST file for `onyo` is intentionally included in git and not
git-ignored because it should be stable, and will act as a canary for
large, unintentional changes.