DOC/BF: fix docstring escapes for Python 3.12
rST requires that string formatting (backticks, emphasis, etc) be
surrounded by whitespace. Thus the following is illegal:

  ``ASSET``s

Escaping the "s" used to be OK, but produces a warning in Python 3.12.

  ``ASSET``\s

The "correct" solution is to escape a space, which (oddly enough) ends
up swallowing the space. This is seriously how rST is meant to be
used. See https://documatt.com/restructuredtext-reference/element/strong-emphasis.html#requires-whitespace-or-punctuation-around

So we move to:

  ``ASSET``\ s

Because that is so much easier to read...