DOC: add docstrings for StoreKeyValuePairs
This is a weird one, since the only consumer of this is actually
Argparse, and it's the one dictating the parameters, etc.

For that reason, I did not document the parameters beyond stating that
they exist. I was going to steal upstream's docstrings, but they don't
have any. At all. Not for Action, nor for anything that Subclasses it.