DOC/RF: docstrings for raise_on_inventory_state(), fsck(), onyo_cat(), onyo_config()
This commit primarily focuses on improving the docstrings for:
- raise_on_inventory_state()
- fsck()
- onyo_cat()
- onyo_config()

Some comments and whitepace are also altered, to improve readability.

There is a small change to onyo_cat() that moves asset validation to be
after printing the asset contents. This does not change the
functionality of onyo_cat(). IMO it's more readable, more accurate, and
removes the risk of anything exploding before printing content (which is
the purpose of onyo_cat()).

works towards #492 #493