`new`: Allow actual paths for `--template`
Changes `new`'s `--template` parameter to accept actual paths rather
than just names. If a path given is relative, it is first looked for
underneath the template dir, maintaining the previous behavior when one
gives a name of template file therein. Only, if that doesn't exist is a
relative path resolved normally (CWD).

Closes #593