Consistently fail on post-argparse option parsing
Allow subcommands to fail on invalid arguments in a way
consistent with argparse, when the constraint can't be (or isn't anyway)
put into the argument specification dict.
Introduces a dedicated exception, that is handled to look like an
argparse error in `main.py`. Meaning: style of error reporting, printing
usage of subcommand and return code 2.

With that, allow `new`'s `-d/--directory` option only once.
(Closes #539)