UI: Allow more flexible questions
This changes the `request_user_response` function to be able to pose
more than just yes-no-questions. Previous behavior remains the default,
so asking a yes-no-question where `UI.yes` would lead to a
non-interactive 'yes' answer.
However, one can now specify the kind of expected answers, what to
return (obviously not just a `bool`) and what is the default answer.
`UI.yes` would now lead to giving the default answer instead, since
`--yes` is technically meant to be non-interactive mode (and needs to
change accordingly eventually).
With that we can ask questions with more than two options to answer
and/or asking for values other than `True` or `False`.
In addition the default answer is assumed when the response is empty
(user just hit enter).