`new`/`edit`: 3- and 4-way questions for editing assets
This allows for a better UX when editing assets. One now gets a 4-way
question when asked whether to accept the edit of an asset, allowing to
either accept the asset (and possibly continue with the next one), edit
it again, skip the asset keeping changes to possible previous assets and
continue editing what assets remain in the list, or abort the command
altogether discarding everything.
In case of a validation error (invalid YAML or something), this becomes
e three-way question, since the asset cannot be accepted as is.