Test for empty keys in inventory operations
As part of a refactoring effort to straighten out validation, rename
the raise_required_key_empty function to reflect it's actually about
empty values and add a function to test for actual keys being empty.

With `add_asset` and `modify_asset` making use of both, the respective
test in `onyo_set` became superfluous and was removed.

(Closes #535)