UX: Don't pollute user facing output
In `main.py` we catch and somehow print/log any exception that bubbles
up. At this point we can and should deal with different exceptions
differently.
A `CalledProcessError` giving the command that was called and
potentially its stdout/stderr and return code in a printed dictionary
isn't usually nice or useful for an onyo user. Normally, we let a
subproccess' stdout/stderr through already - that part would be a
duplicate. Plus, it's not nice to read and a user doesn't normally nned
to know what exactly we called internally.
This was particularly apparent w/ `onyo config` where we aim to mimic
the behavior of underlying git-config. If one queries for a config that
doesn't exist, git-config remains silent and returns 1. We don't want to
see an error there.
Hence, only relay the return code and log the actual exception at debug
level.