Account for repo versions
This is starting to account for different repository versions.
Since the asset name config was changed and that's a functional part,
the repository version also changed.
With this patch, we detect unknown repository versions - too old (no
version recorded) or newer than the onyo version running and fail
accordingly.
In addition there is a bit of a shim code to translate the changed
config names when onyo encounters a version 1 repository.
Eventually, such code should be refactored so that necessary conversions
are centrally defined and we should have a way to actually migrate
repositories.

(Closes #441)