Fix error when CWD doesn't exist
CWD is queried already when the `argparse` parser is set up. Hence,
an exception occuring during that (like running from a shell that still
has a CWD that doesn't exist anymore) would not go through the generic
handler for exception coming from onyo command implementations.
As a consequence, there was no output format and a full traceback was
dumped to the user, which may be way more confusing/scary than it needs
to be.