`get`: More flexible sort option
This changes the sorting options for `onyo_get` to allow specifying
multiple keys to sort by, in what order and ascending/descending order
per key.

In order to have a CLI that allows to specify ascending/descending via
the formally different arguments `-s` and `-S` respectively,
while also maintaining order of the keys specified, something of a hack
was needed, where a dedicated action for `argparse` is implemented, that
ignores the original `dest` in the namespace returned by argparse, and
puts keys specified via either argument into the same namespace
attribute.

(Closes #507)