Colorize diff output
While this works rendering diff output colorized similar to `git`,
this solution isn't quite there yet and should be integrated in `UI`.
However, this involves a general solution to printing via `rich` where
we make sure our tests can deal with that and don't confuse `rich` as to
whether it's writing to a terminal when we capture output for testing.
See also #542.

(Closes #581)