TMP: Workaround YAML representation issue
When reading from (edited) files, ruamel delivers a dict-like
(`CommentedMap`). When accessed like a dict something like `serial:
00123` would show up as `{'serial': 123}`. However, under the hood
ruamel retains the information how it was read (in roundtrip mode) and
consequently dumps it as `serial: 00123`, which is what we want.

The issue is, that when using these dicts "normally", we end up looking
at the "wrong" value - in this case an integer `123` - which messes up
dumping into the asset name as well as dict comparisons when trying to
find out whether there even was a modification.

This patch introduces a dict-like wrapper that delivers the
representation ruamel would use for dumping rather than the "technical
value" of that object.
The patch is considered temporary as it is incomplete (not using this
mechanic for matching assets in `onyo get` for example) and should be
integrated with similar efforts, most notably accounting for comments
when comparing assets and implementing full nested dictionary support.