BF: make sorting case sensitive
Previously, sorting was case insensitive, so that the words would
actually be grouped together. However, within a word, it was not sorted.

GROUPLETTERS does not do what we want, but using LOCALE does (at least
in 'en_US.UTF-8').

For now, I am choosing to use whatever the user has configured on their
OS. That makes sense to me, respecting their choice. If this ends up
biting us, then we'll simply hardcode it or make it a configurable
option in `.onyo/config`.

fix #651