RF: simplify template tab completion
I don't quite know what I was trying to achieve here before, but with
how we're currently using Onyo, this seems needlessly complicated.

Rarely, people report seeing the following when they first use
`onyo new --template`:

> onyo new -t (eval):1: command not found: _template_dir

I do not know what causes that, but it so far is unique to
_template_dir.

These changes simplify it to work like all the others.