BF: add explicit `---` document start to YAML
It seems that ruamel does not retain the DocumentStartToken or anything
before it in its comment map, even in round-trip mode. However, one can
explicitly tell ruamel to generate the `---`, which this change does.

A limitation of this approach is that changes to the DocumentStartToken
(and earlier) will not be detected nor diffed by Onyo. Sshort of
reverting to file-based diffs, this is not practical. So we accept it
for what it is.

close #659