Implement wrapper class for dictionary support
This implements a wrapper class for asset dictionaries, that allows to
access keys in nested dictionaries using a period. In other words: This
class gives a dict-like view on a flattened dict. Consequently, the
behavior of iterating changes as well, providing only the flattened
view. The underlying (non-flat) dictionary remains available via
``.data``.
Therefore, this wrapper should allow for implementing dictionary support
with little changes to existing code, since there are only a few entry
points where dicts would need to be wrapped that way, while everything
else just uses this like a regular ``dict``.

The ultimate goal here is, to be able to provide these
"hierarchical keys" to the CLI (and allow its use in configs like
``onyo.assets.name-format``). Adressing keys of nested dictionaries that
way is relatively simple and straight-forward. As opposed to a solution
that amounts to learning a "new language" from a user's perspective.
With that implementation there's a price: No literal periods in key
names allowed. If needed, this implementation could be changed to first
check for literal dots in key names and have the now implemented
behavior be a fallback instead. This may lead to ambiguity when
setting/creating new keys, though. So, let's deal with that complication
only when we figure we actually need it.