Adjust tests for dictionary support
Test the support of (nested) dictionaries in assets' YAML content.
For the purpose of testing this is using the INM-7 usecase where we want
``model`` to be a dict w/ subkeys like ``name`` (used in the asset name
configuration) and possibly ``year``, ``series`` or similar,
possibly manufacturer depend properties.

Note, that the name config is only changed for testing here. This patch
does not change onyo's default config.