`tree:`: directories only
This implements a directories only mode for `onyo tree`.
Breaking change: Rename the existing `directories` argument for the
python interface to `paths` in order to be less confusing. This makes
sense in python as that argument is an interable of `Path` objects and
it's irrelevant to the CLI, since it's a positional argument - so, no
confusing option names.

In addition this fixes the command's help referring to a metavar for the
positional argument that doesn't exist.
Lastly, a slight performance improvement for `onyo tree` that avoids
`stat`ing the same path several times to tell whether it's a directory.

Closes #677