Make commit messages more flexible
This patch changes the behavior of the `--message` switch for committing
onyo commands to append the given message to the auto-generated subject
line. In addition it introduces a `--no-auto-message` switch that allows
to explicitly disable the auto-generation of such a subject line on a
per-call basis. For the python interface the name of this switch changes
to `auto_message` - no need to deal with double negation.
The default behavior can be configured via `onyo.commit.auto-message`
and this variable is written by `onyo init` as `true`.

Closes #685