NF: add InvalidAssetError and use in `onyo_cat()`
I suspect `OnyoInvalidRepoError` was used because that's what `fsck` is
throwing. Much it's really misleading when we know the repository is
fine, we're not doing a repo check, and we happen to encounter an
invalid asset.

I imagine this will be useful when we engage in validation, and need to
catch such events.