RF/DOC: normalize `onyo history`
This normalizes the `onyo history` command to gain (and call) an
`onyo_history()` function and follow other common idioms.

Error codes now bubble up the same as all other shell calls.

`get_history_cmd()` has been moved from command_utils.py and adjusted to
be a private helper function for `onyo_history()`.

The `interactive` parameter has been enhanced to allow forcing
interactive on (`True`) or off (`False`), and auto-detection (`None`).
Previously `True` meant auto-detection and one could not force on
interactive mode.

The `onyo history` command is still unable to force interactive mode to
be on. Such functionality is not considered to be valuable, but can be
added trivially if it is later deemed desirable.

The `onyo history` command no longer checks if the target path exists.
That's considered the responsibility of the configured history command.

I made a serious effort to allow multiple paths to be passed to `onyo
history`, and allow the configured program to throw an error if it can
only accept one path.  Both `tig` and `git log` allow multiple paths,
but neither allow it with the `--follow` flag enabled, which is simply
crucial for Onyo. Files move, a lot. So I accepted it as a limitation.

close #494