(Mostly) ignore yaml implicit typing
This is disabling yaml implicit typing except for `null`.
Everything, keys and values, are read as literal strings be it from CLI
or from actual YAML files, allowing users to ignore the weirdness that
is YAML's implicit typing, like `003_5` being the integer 35, while not
impeding the abilities of onyo. There's no functionality that would
actually use types (Onyo isn't doing arithmetic on values or something
of the sort).

Explicit typing, while not currently used, is untouched, though.
`!!int` is read as an actual integer and a python integer is written
that way.

Closes #696