TST: add Faker fixture ("fake")
Will be used for upcoming benchmark tests, but will also be useful
across tests in general.

This also makes them available across scopes. Unsure about the utility
of this, given the large number of bugs I've encountered with fixture
scope and parametrization, but I wanted this in the repo history so we
can refer to it.

fix #254