Introduction of Item class
This patch is a first step towards implementing new pseudo-keys, key
namespaces, and key aliases (github issues #688, #689).
The ``Item``-class is a dict-like that provides pseudo-keys for anything
that could be tracked by onyo (YAML-files and directories).
``Item`` is therefore replacing ``DotNotationWrapper`` except for places
where the (test-)code does not want pseudo-keys (yet).
Its ability to represent other things than just assets (directories,
templates, etc.) is not yet utilized in this commit. The aim is for
``Item`` to be the one datastructure that any ``Inventory`` level code
is acting upon and passing around instead of ``dict`` and ``Path``.

This commit is just introducing this class and replaces internal usage
of the 'path', 'directory', and 'is_asset_directory' keys. The two
former become aliases to `onyo.path.relative` and `onyo.path.parent`
instead and are still partially used in tests for that reason.

The 'is_asset_directory' key is retired. This is now covered by the
'onyo.is.directory' and 'onyo.is.asset' keys. In most cases, that means
to use 'onyo.is.directory' instead, since the command/function context
already determines that this is about an asset (so, no need to specify
this redundantly via 'onyo.is.asset'.