BF: Don't end up with unset pseudo-keys after edit
When reading an asset's content from the temporary file for editing
(``onyo new --edit``/``onyo edit``), do not remove the `Item`'s
pseudo-keys. This happened, because we overwrite the ``.data`` attribute
of the ``Item`` directly. That's necessary for ruamel's comment
roundtrip, but we need to put the pseudo-key definitions back in
afterwards.
Failing to do so, resulted in unset pseudo-keys, which in case of
``onyo.path.relative`` then generated commit messages with ``None`` as
the asset path.

Closes #739