NF: DotNotationWrapper support ingesting dotnotation
This adds a `pristine_original` parameter that defaults to `True` (the
original behavior).

It can be set to `False` to all the incoming dict to be interpreted for
dots in its key names. This is especially useful when ingesting (for
example) a TSV with key names already in dot notation. Where there is
not original YAML to wish to preserve, etc.