RF: Item-ize onyo_mkdir() and inventory.add_directory()
This takes to approach of: minimal changes to make things work in areas
of the code that are not ready to be flipped fully over to Item yet.

This will lead to a bit of churn, but is the only rational way to
convert the codebase incrementally rather than all at once.