Remove types=templates from querying
This removes templates as a type from `onyo_get` and the underlying
functions again and enables the `onyo.is.asset` and `onyo.is.directory`
pseudo-keys for templates.
This is done in preparation of allowing (asset-)directory templates,
which requires that templates can specify these properties.
In order to query templates, one now has to explicitly pass a template
dir to `get`'s `include` parameter and the types to query can be
specified just like when querying the actual inventory.