RF: cache OnyoRepo.get_config()
This introduces a simple cache at the OnyoRepo layer that is cleared on
`commit()` or `set_config()`. It can be cleared manually with
`clear_cache()`.

GitRepo is left untouched and is available for cache-less queries.

The motivation is the significant overhead of querying
`onyo.assets.name-format`, which is used on a per-asset basis by `set`,
`new` and (for now) `get`.

Initially, I attempted to introduce a per-file file cache at the GitRepo
layer, and track the mtime of files that would be queried. However, in
the case of a bare `git config` call, the number of locations is quite a
few and are modified by environmental variables. These environmental
variables would need to be resolved and tracked as well for changes.

CLI users will not have any issues managing the cache (and the new
behavior is arguably an improvement, as Onyo's behavior can no longer
change mid-flight if the system config changes in parallel). For users
of the API, they only encounter issues if they explicitly avoid
`OnyoRepo.set_config()`; in which case `clear_cache()` solves their
problems.