Add multidoc YAML reading
Add a generic multidoc YAML reader. This is done in preparation of
multidoc templates and YAML inputs for `onyo_new()`.
In addition this allows asset files to contain multiple documents,
where any documents past the first one are ignored by onyo, allowing for
a kind of sidecar documents in asset files.