Proof of concept show <-> new
This is a proof-of-concept setup, where `whosyourdaddy()` produces YAML
multidoc output from templates/assets/directories in the inventory,
suitable for the envisioned `onyo show`.

The other way around, `iamyourfather()` is reading that in and producing
a list of (to-be created) `Item`s that could be passed to
`Inventory.add_asset()`/`Inventory.add_directory()`.